[OPR-405]: Changes AbstractProcedure back to Procedure
